part_2/requirements.txt
celery[redis]
psycopg2-binary
requests
python-dotenv
